pyproject.toml
[project]
name = "kdtree"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "scikit-image>=0.25.2",
    "scipy>=1.15.3",
    "tqdm>=4.67.1",
]
